+++ spec/models/user_activity_spec.rb
@@ -297,6 +297,46 @@ describe UserActivity do
         end
       end
     end
+
+    context 'with per-channel sync settings' do
+      before do
+        allow_any_instance_of(Team).to receive(:slack_channels).and_return([{ 'id' => 'channel_id' }])
+        allow_any_instance_of(User).to receive(:user_deleted?).and_return(false)
+        allow_any_instance_of(User).to receive(:user_in_channel?).and_return(true)
+      end
+
+      context 'when sync is disabled for the channel' do
+        before { user.set_user_channel!('channel_id', 'general', sync_activities: false) }
+
+        it 'does not post to the channel' do
+          expect(user.team.slack_client).not_to receive(:chat_postMessage)
+          expect(activity.brag!).to eq([])
+        end
+
+        it 'still sets bragged_at' do
+          activity.brag!
+          expect(activity.reload.bragged_at).not_to be_nil
+        end
+      end
+
+      context 'when sync is enabled for the channel' do
+        before { user.set_user_channel!('channel_id', 'general', sync_activities: true) }
+
+        it 'posts to the channel' do
+          expect(user.team.slack_client).to receive(:chat_postMessage).and_return('ts' => 1)
+          expect(activity.brag!).to eq([{ ts: 1, channel: 'channel_id' }])
+        end
+      end
+
+      context 'when no per-channel setting exists and global sync is disabled' do
+        before { user.update_attributes!(sync_activities: false) }
+
+        it 'does not post to the channel' do
+          expect(user.team.slack_client).not_to receive(:chat_postMessage)
+          expect(activity.brag!).to eq([])
+        end
+      end
+    end
   end
 
   describe '#display_title_s' do